Refactor of SearchHeader and SearchFooter to use only one component with different renders
